requirements.txt
langchain-google-genai
langchain-community
langchain
pypdf
Pillow
python-dotenv
streamlit
